recommendations/requirements.txt
exceptiongroup==1.1.2
grpc-interceptor==0.15.2
grpcio==1.56.0
grpcio-tools==1.56.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
protobuf==4.23.4
pytest==7.4.0
tomli==2.0.1
